Trust in the Police in Belgium (European Social Survey, Round 5)

This is a sample data set cobbled from the fifth round of European Social Survey data for Belgium. It offers a means to do a basic replication of some of Chapter 5 of The SAGE Handbook of Regression Analysis and Causal Inference.

ESSBE5

Format

A data frame with 1704 observations on the following 10 variables.

essround

a numeric for the ESS round

edition

a character for the edition number of the fifth round

idno

a numeric id number

cntry

a character vector for the country (i.e. Belgium, or BE)

trstplc

a numeric vector for trust in the police on an 11-point scale. Higher values indicate more trust. 0 = "no trust at all". 10 = "complete trust"

agea

a numeric vector for the respondent's age

female

a numeric vector for whether the respondent is a woman or not.

eduyrs

a numeric vector for years of education.

hincfel

a numeric vector for the respondent's feeling about their household income. 1 = "living comfortably", 2 = "coping on present income", 3 = "difficult on present income", 4 = "very difficult on present income"

plcpvcr

a numeric vector for how successful police are at preventing crimes in a country on an 11-point scale. 0 = "extremely unsuccessful". 10 = "extremely successful."

Source

European Social Survey (Round 5)

Details

See Chapter 5 of The SAGE Handbook of Regression Analysis and Causal Inference for more information.