European Union Convergence (Maastricht, SGP) Criteria

A data set on fiscal indicators communicating performance against convergence criteria outlined in the Maastricht Treaty and Stability and Growth Pact (SGP).

Usage

EUCC

Format

A data frame with 30 observations on the following 6 variables.

geo

a Eurostat code

geoname

an English country name or other identifier for the Eurostat code

agg

a dummy variable indicating whether the geo entry is some kind of aggregate of countries

gdp

GDP in millions of Euros

debt

general government gross debt, nominal, in millions of Euros

bb

budget balance (i.e. deficit/surplus), nominal, in millions of Euros

Details

Data come from Eurostat. Codes for GDP, debt, fiscal balance measures are nama_10_gdp, teina225 and teina200, respectively. na_item for the GDP observations is B1GQ. Observations are for 2025. Some GDP numbers are provisional but ultimately fine for this purpose.

EU criteria oblige member states to not have a budget balance lower than 3% of GDP (i.e. -3%, in this measure) and to not have debt exceed 60% of GDP.