School Expenditures and Test Scores for 50 States, 1994-95

A data set for a canonical case of a Simpson's paradox, useful for in-class instruction on the topic.

Guber99

Format

A data frame with 50 observations on the following 8 variables.

state

a character vector for the state

expendpp

a numeric vector for the current expenditure per pupil in average daily attendance in public elementary and secondary schools, 1994-95 (in thousands of dollars)

ptratio

a numeric vector for the average pupil/teacher ratio in public elementary and secondary schools, Fall 1994

tsalary

a numeric vector for the estimated average annual salary of teachers in public elementary and secondary schools, 1994-95 (in thousands of dollars)

perctakers

a numeric vector for the percentage of all eligible students taking the SAT, 1994-95

verbal

a numeric vector for the average verbal SAT score, 1994-95

math

a numeric vector for the average math SAT score, 1994-95

total

a numeric vector for the average total SAT score, 1994-95

References

Guber, Deborah Lynne. 1999. "Getting What You Pay For: The Debate Over Equity in Public School Expenditures." Journal of Statistics Education 7(2).