Partisan Politics in the Global Economy

A data set on government spending in select rich countries as a function of trade/GDP, financial openness, and the state-year-level engagement in trade unions. The data offer a means to quasi-replicate Garrett's (1998) argument about left-wing governments' ability to stem the tide of globalization's effect on decreased government spending.

Usage

PPGE

Format

A data frame with the following 9 variables.

country

a character vector for the country

iso3c

a character vector for the three-character country ISO code

year

the year

govtspendgdp

total government spending over GDP

tradegdp

the volume of trade over GDP

kaopen

an index measuring a country's degree of capital account openness

ka_open

an alternate index measuring a country's degree of capital account openness, normalized to be between 0 and 1

v2catrauni

an estimate of a country's engagement in independent trade unions, generated by way of a Bayesian item response model

v2catrauni_ord

an estimate of a country's engagement in independent trade unions, on ordinal scale. See details.

Details

The data are an unbalanced panel because of data missingness primarily affecting Switzerland (which would only appear in the panel in earnest starting in the mid-1990s). The Netherlands has some missing data in the mid-1970s. Spain and Portugal appear at the start of the panel, though the transition to democracy for both wouldn't start until 1974/1975. The data also have some obvious COVID weirdness for 2020. Perhaps an honest re-assessment of Garrett (1998) may want to drop Switzerland altogether, ignore 2020, and lop a few years off the Spanish and Portuguese panel. What you do with the Netherlands is up to you.

Briefly: the government spending/GDP data come from the International Monetary Fund. The trade/GDP data come from the World Bank's API. The financial openness indicators come by way of the Chinn-Ito index. The engagement in trade unions data are from the Varieties of Democracy project. The ordinal measure of the trade union estimates communicate what percentage of the population is active in independent trade unions. Values include 0) virtually no one 1) a small share of the population (less than 5 moderate share of the population (about 5 to 15 the population (about 16 (about 26

References

Coppedge, Michael, John Gerring, Carl Henrik Knutsen, Staffan I. Lindberg, Jan Teorell, Nazifa Alizada, David Altman, Michael Bernhard, Agnes Cornell, M. Steven Fish, Lisa Gastaldi, Haakon Gjerløw, Adam Glynn, Sandra Grahn, Allen Hicken, Garry Hindle, Nina Ilchenko, Katrin Kinzelbach, Joshua Krusell, Kyle L. Marquardt, Kelly McMann, Valeriya Mechkova, Juraj Medzihorsky, Pamela Paxton, Daniel Pemstein, Josefine Pernes, Oskar Rydén, Johannes von Römer, Brigitte Seim, Rachel Sigman, Svend-Erik Skaaning, Jeffrey Staton, Aksel Sundström, Eitan Tzelgov, Yi-ting Wang, Tore Wig, Steven Wilson and Daniel Ziblatt. 2022. "V-Dem [Country-Year/Country-Date] Dataset v12" Varieties of Democracy (V-Dem) Project. doi: 10.23696/vdemds22

Chinn, Menzie D. and Hiro Ito. 2006. "What Matters for Financial Development? Capital Controls, Institutions, and Interactions." *Journal of Development Economics* 81(1): 163--192.

Garrett, Geoffrey. 1998. *Partisan Politics in the Global Economy* New York, NY: Cambridge University Press.