Urban Heat Islands and Green Land Use in European Cities

A data set on urban heat island magnitude in European cities.

Usage

eurouhi

Format

A data frame with 285 observations on the following 14 variables.

city

a city identifier

pf

the planning tradition/family, one of "British", "Central", "Mediterranean", "New member", or "Nordic"

uhimag

a measure of the urban heat island magnitude

long

the longitude of the area in question

lat

the latitude of the area in question

mindistcoast

the minimum distance to the coast, in kilometers

aaperc

the percentage of the area that's artificial, as a proportion

agperc

the percentage of the area that's agricultural, as a proportion

forperc

the percentage of the area that's forest, as a proportion

ugperc

the percentage of the area that's urban green, as a proportion

avgpa

the mean patch area, in hectares

edgedens

edge density in meter/square meters

lpi

the largest patch index, as a percentage (proportion)

plc

the proportional landscape core

Details

Data are able to replicate Figures 2, 3, and 4 in Nastran et al. (2019). Consult that article for more information.

References

Nastran, Mojca, Milan Kobal, and Klemen Eler. 2019. "Urban Heat Islands in Relation to Green Land Use in European Cities." Urban Forestry & Urban Greening 37: 33-41.