The Effect of Urban Parks on the Thermal Environment in Shanghai

A data set on urban parks in Shanghai and their effect on temperature.

Usage

shuparks

Format

A data frame with 24 observations on the following 10 variables.

group

a grouping category for park size, either 'Super Large', 'Large', 'Medium', or 'Small'

name

the name of the park

ha

park area, in hectares

areap

percentage of the area of each park in the built-up part of Shanghai

gc

the percentage of the trees or grass area in each park divided by the park area

func

the supposed function of each park, per the Shanghai Landscape and City Appearance Administrative Bureau and Forestry Bureau

mcd15

the maximum cooling distance, in meters, in 2015

mci15

the maximum cooling intensity, in Celsius, in 2015

mcd20

the maximum cooling distance, in meters, in 2020

mci20

the maximum cooling intensity, in Celsius, in 2020

Details

Data come from Table 1 and Table 5 in Wang et al. (2022). Read that article for more technical details.

References

Wang, Tian, Gui Tu, Bo Min, Zuzheng Li, Xiaofang Li, and Qingxiang You. 2022. "The Mitigation Effect of Park Landscape on Thermal Environment in Shanghai City Based on Remote Sensing Retrieval Method." International Journal of Environmental Research and Public Health 19(5): 2949.