Thermometer Ratings for Donald Trump and Barack Obama

A data set on thermometer ratings for Donald Trump and Barack Obama in 2020. I use these data for in-class illustration of central limit theorem. Basically: the sampling distribution of a population is normal, even if the underlying population is decidedly not.

Usage

therms

Format

A data frame with 3080 observations on the following 2 variables.

fttrump1

a thermometer rating for Donald Trump

ftobama1

a thermometer rating for Barack Obama

Source

American National Election Studies (ANES) Exploratory Testing Survey (ETS)

Details

The survey period was April 10-18, 2020 and was done entirely online. Thermometer ratings are on a 0 to 100 scale, where higher values indicate more "warmth".