South Carolina County GOP/Democratic Primary Data, 2016

County-level data on vote share and various background/demographic information for the 2016 South Carolina GOP/Democratic primaries.

SCP16

Format

A data frame with 46 observations on the following 15 variables.

county

the county

clinton

Hillary Clinton's county-level vote share in the 2016 party primary

sanders

Bernie Sanders' county-level vote share in the 2016 party primary

trump

Donald Trump's county-level vote share in the 2016 party primary

cruz

Ted Cruz' county-level vote share in the 2016 party primary

rubio

Marco Rubio's county-level vote share in the 2016 party primary

percapinc

A county-level estimate for per capita income

medhouseinc

A county-level estimate for the median household income

medfaminc

A county-level estimate for the median family income

illiteracy

An estimate of the percent of the county lacking "basic" prose literacy skills

perblack

Percentage of the county that is black

population

An estimate of the county-level population

romneyshare2012

Mitt Romney's vote share at the county-level from the 2012 general election

perhsgrad

Percentage of the county whose residents 25 years and older have at least a high school education

unemployment

Unemployment rate for the county for January 2016

Details

The illiteracy estimate comes from a Department of Education report from 2003. The unemployment rate data come from the Bureau of Labor Statistics. A Github repository contains more information: https://github.com/svmiller/sc-primary-2016.