Illiteracy in the Population 10 Years Old and Over, 1930

This is perhaps the canonical data set for illustrating the ecological fallacy.

illiteracy30

Format

A data frame with 40 observations on the following 11 variables.

state

a character for the state

pop

a numeric vector for the total population

pop_il

a numeric vector for the total population that is illiterate

nwhite

a numeric vector for the total native white population

nwhite_il

a numeric vector for the total native white population that is illiterate

fpwhite

a numeric vector for the total white population with "foreign or mixed parentage"

fpwhite_il

a numeric vector for the total white population with "foreign or mixed parentage" that is illiterate

fbwhite

a numeric vector for the total foreign-born white population

fbwhite_il

a numeric vector for the total foreign-born white population that is illiterate

black

a numeric vector for the total black population.

black_il

a numeric vector for the total black population that is illiterate

Source

U.S. Census Bureau (1933). Fifteenth Census of the United States: 1930. Population, Volume II.

Details

All population totals reflect those 10 years or older. The 1930 Census (along with Robinson (1950)) uses "negro" in lieu of black, but the variable names here eschew that older label. Note that some states are not yet states in the 1930 Census.

References

Grotenhuis, Manfred Te, Rob Eisinga, and SV Subramanian. 2011. "Robinson's Ecological Correlations and the Behavior of Individuals: methodological corrections." Internatoinal Journal of Epidemiology 40(4): 1123-25.

Robinson, WS. 1950. "Ecological Correlations and the Behavior of Individuals." American Sociological Review 15(3): 351--57.